#778: Align CALM Hub and CALM Visualizer
