+++ calm-hub/src/main/webapp/package.json
@@ -7,9 +7,18 @@
     "start": "vite",
     "build": "vite build",
     "serve": "vite preview",
-    "prod": "vite build && mkdir -p ../resources/META-INF/resources && rsync -a build/* ../resources/META-INF/resources"
+    "prod": "vite build && mkdir -p ../resources/META-INF/resources && rsync -a build/* ../resources/META-INF/resources",
+    "test": "vitest",
+    "lint": "eslint .",        
+    "format": "prettier --write ."
   },
   "dependencies": {
+    "cytoscape": "^3.30.3",
+    "cytoscape-cola": "^2.5.1",
+    "cytoscape-dagre": "^2.5.0",
+    "cytoscape-expand-collapse": "^4.1.1",
+    "cytoscape-node-edge-html-label": "^1.0.6",
+    "react-icons": "^5.3.0",
     "@patternfly/patternfly": "^4.215.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
@@ -22,6 +31,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-json-view-lite": "^2.0.1",
+    "react-router-dom": "^7.2.0",
     "typescript": "^4.9.5",
     "vite": "^6.0.7",
     "vite-plugin-svgr": "^4.3.0",
@@ -55,12 +65,30 @@
   },
   "devDependencies": {
     "autoprefixer": "^10.4.20",
-    "eslint": "^8.57.1",
+    "eslint": "^9.14.0",
     "eslint-plugin-react": "^7.37.2",
     "eslint-plugin-react-hooks": "^5.0.0",
     "jsdom": "^25.0.1",
     "npm-force-resolutions": "^0.0.10",
     "postcss": "^8.4.49",
-    "tailwindcss": "^3.4.17"
+    "tailwindcss": "^3.4.17",
+    "@eslint/js": "^9.14.0",
+    "@tailwindcss/typography": "^0.5.16",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.1.0",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/cytoscape": "^3.21.8",
+    "@types/file-saver": "^2.0.7",
+    "@types/react": "^18.3.10",
+    "@types/react-dom": "^18.3.0",
+    "@vitejs/plugin-react": "^4.3.2",
+    "daisyui": "^4.12.14",
+    "eslint-plugin-react-refresh": "^0.4.12",
+    "globals": "^15.12.0",
+    "prettier": "^3.3.3",
+    "typescript": "^5.5.3",
+    "typescript-eslint": "^8.14.0",
+    "vite": "^5.4.8",
+    "vitest": "^2.1.8"
   }
 }
