+++ calm-hub/src/main/webapp/postcss.config.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+  plugins: {
+      tailwindcss: {},
+      autoprefixer: {},
+  },
+};
