+++ calm-hub/src/main/webapp/src/components/navbar/Navbar.css
@@ -1,0 +1,3 @@
+.navbar {
+ padding: 0rem !important;
+}
