+++ calm-hub/src/main/webapp/src/hub/components/value-table.tsx
