+++ calm-hub/src/main/webapp/src/index.tsx
@@ -4,6 +4,7 @@ import './index.css';
 import '@patternfly/patternfly/patternfly.css';
 import App from './App.js';
 
+
 const root = ReactDOM.createRoot(document.getElementById('root')!);
 root.render(
     <React.StrictMode>
