+++ calm-hub/src/main/webapp/src/visualizer/components/sidebar/Sidebar.tsx
@@ -1,5 +1,5 @@
 import { IoCloseOutline } from 'react-icons/io5';
-import { Edge, Node } from '../cytoscape-renderer/CytoscapeRenderer';
+import { Edge, Node } from '../cytoscape-renderer/CytoscapeRenderer.js';
 
 interface SidebarProps {
     selectedData: Node['data'] | Edge['data'];
@@ -20,9 +20,9 @@ function Sidebar({ selectedData, closeSidebar }: SidebarProps) {
     const isCALMEdge = isCALMEdgeData(selectedData);
 
     return (
-        <div className="fixed right-0 h-full w-80 bg-gray-100 shadow-lg">
+        <div className="fixed right-0 h-full w-80 bg-base-300 shadow-lg">
             <label htmlFor="node-details" className="drawer-overlay" onClick={closeSidebar}></label>
-            <div className="menu bg-base-200 text-base-content min-h-full w-80 p-4">
+            <div className="menu bg-base-300 text-base-content min-h-full w-80 p-4">
                 <div className="flex justify-end">
                     <button
                         aria-label="close-sidebar"
