+++ calm-hub/src/main/webapp/src/visualizer/tests/Drawer.test.tsx
@@ -2,6 +2,7 @@ import { describe, expect, it } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import Drawer from '../components/drawer/Drawer';
 
+
 describe('Drawer', () => {
     it('should render Drawer', () => {
         render(
