--- calm-visualizer/.prettierrc
@@ -1,9 +1,0 @@
-{
-    "trailingComma": "es5",
-    "tabWidth": 4,
-    "semi": true,
-    "singleQuote": true,
-    "bracketSpacing": true,
-    "endOfLine": "auto",
-    "printWidth": 100
-}
