--- calm-visualizer/README.md
@@ -1,45 +1,0 @@
-## Calm Visualizer
-
-Calm Visualizer is a tool designed to help visualize and manage software architecture through a UI.
-
-### Features
-
-- Visualize uploaded architecure JSON file using cytoscape JS
-- Toggle connection and node descriptions
-
-### Installation
-
-To install Calm Visualizer, clone the repository and install the dependencies:
-
-```sh
-cd calm-visualizer
-npm install
-```
-
-### Usage
-
-To start the UI, run:
-
-```sh
-npm run dev
-```
-
-This will start the UI and you can access the application at `http://localhost:5174`.
-
-### Building for Production
-
-To build the application for production, run:
-
-```sh
-npm run build
-```
-
-This will create a `dist` directory with the production build of the application.
-
-## License
-
-Copyright 2024 FINOS
-
-Distributed under the [Apache License, Version 2.0](http://www.apache.org/licenses/LICENSE-2.0).
-
-SPDX-License-Identifier: [Apache-2.0](https://spdx.org/licenses/Apache-2.0)
