--- calm-visualizer/package.json
@@ -1,52 +1,0 @@
-{
-    "name": "calm-visualizer",
-    "private": true,
-    "version": "0.0.0",
-    "type": "module",
-    "scripts": {
-        "dev": "vite",
-        "build": "vite build",
-        "lint": "eslint .",
-        "format": "prettier --write .",
-        "preview": "vite preview",
-        "test": "vitest"
-    },
-    "dependencies": {
-        "cytoscape": "^3.30.3",
-        "cytoscape-cola": "^2.5.1",
-        "cytoscape-dagre": "^2.5.0",
-        "cytoscape-expand-collapse": "^4.1.1",
-        "cytoscape-node-edge-html-label": "^1.0.6",
-        "file-saver": "^2.0.5",
-        "react": "^18.3.1",
-        "react-dom": "^18.3.1",
-        "react-icons": "^5.3.0"
-    },
-    "devDependencies": {
-        "@eslint/js": "^9.14.0",
-        "@tailwindcss/typography": "^0.5.16",
-        "@testing-library/jest-dom": "^6.6.3",
-        "@testing-library/react": "^16.1.0",
-        "@testing-library/user-event": "^14.5.2",
-        "@types/cytoscape": "^3.21.8",
-        "@types/file-saver": "^2.0.7",
-        "@types/react": "^18.3.10",
-        "@types/react-dom": "^18.3.0",
-        "@vitejs/plugin-react": "^4.3.2",
-        "autoprefixer": "^10.4.20",
-        "daisyui": "^4.12.14",
-        "eslint": "^9.14.0",
-        "eslint-plugin-react": "^7.37.2",
-        "eslint-plugin-react-hooks": "^5.1.0-rc.0",
-        "eslint-plugin-react-refresh": "^0.4.12",
-        "globals": "^15.12.0",
-        "jsdom": "^25.0.1",
-        "postcss": "^8.4.49",
-        "prettier": "^3.3.3",
-        "tailwindcss": "^3.4.14",
-        "typescript": "^5.5.3",
-        "typescript-eslint": "^8.14.0",
-        "vite": "^5.4.8",
-        "vitest": "^2.1.8"
-    }
-}
