--- calm-visualizer/src/index.css
@@ -1,8 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-:root {
-    position: relative;
-    font-family: Arial, Helvetica, sans-serif;
-}
