--- calm-visualizer/tsconfig.json
@@ -1,7 +1,0 @@
-{
-    "compilerOptions": {
-        "esModuleInterop": true
-    },
-    "files": [],
-    "references": [{ "path": "./tsconfig.app.json" }, { "path": "./tsconfig.node.json" }]
-}
